third_party/WebKit/LayoutTests/http/tests/dnd/drop-file-on-webkitdropzone.html

<!doctype html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
#dropzone {
  display: block;
  border: 1px solid black;
  width: 200px;
  height: 200px;
}
</style>

<p>
  Please download <a download href="resources/dragged-file.txt">this file</a>,
  and drag it into the box below.
</p>

<div id="dropzone" webkitdropzone="copy file:text/plain">
  Drop Here
</div>

<script>

async_test(t => {
  const dropZone = document.querySelector('#dropzone');
  dropZone.ondrop = t.step_func_done(event => {
    event.preventDefault();  // Needed to prevent drop navigation.

    const dataTransfer = event.dataTransfer;
    assert_equals(dataTransfer.items.length, 1);
    const item = dataTransfer.items[0];
    assert_equals(item.kind, 'file');
    assert_equals(item.type, 'text/plain');  // application/x-moz-file in FF
    const file = item.getAsFile();
    assert_equals(file.name, 'dragged-file.txt');
    assert_equals(file.type, 'text/plain');
  });

  const clientRect = dropZone.getBoundingClientRect();
  if (window.eventSender) {
    eventSender.beginDragWithFiles(['resources/dragged-file.txt']);
    const centerX = (clientRect.left + clientRect.right) / 2;
    const centerY = (clientRect.top + clientRect.bottom) / 2;
    eventSender.mouseMoveTo(centerX, centerY);
    eventSender.mouseUp();
  }
}, 'webkitdropzone element receives file drop on page served from HTTP');

</script>
